feat: replace FITSIO with FITSFiles
This seems to work fine with the existing test suite, but some more
examples should be included and more extensive testing done.
Additionally, XSPECModels has a table model wrapper that reads in FITS
files that still needs to be converted to FITSFiles, and checked.

There are also currently some bugs with reading in numeric types in
FITSFiles 0.2.0 that need to be fixed before this can be merged into
main.